Extract shared session infrastructure into AbstractUpgradeStrategy
Move Maven session management (getSession, createMaven4Session,
TransporterFactoryConfig, temp directory helpers, buildEffectiveModel)
from PluginUpgradeStrategy into AbstractUpgradeStrategy to enable
reuse across strategies.
