+++ impl/maven-cli/src/main/java/org/apache/maven/cling/invoker/mvnup/goals/AbstractUpgradeStrategy.java
@@ -18,9 +18,13 @@
  */
 package org.apache.maven.cling.invoker.mvnup.goals;
 
+import java.io.File;
+import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
@@ -28,8 +32,27 @@ import eu.maveniverse.domtrip.Document;
 import eu.maveniverse.domtrip.Element;
 import eu.maveniverse.domtrip.maven.Coordinates;
 import eu.maveniverse.domtrip.maven.MavenPomElements;
+import org.apache.maven.api.RemoteRepository;
+import org.apache.maven.api.Session;
 import org.apache.maven.api.cli.mvnup.UpgradeOptions;
+import org.apache.maven.api.di.Named;
+import org.apache.maven.api.di.Provides;
+import org.apache.maven.api.model.Repository;
+import org.apache.maven.api.model.RepositoryPolicy;
+import org.apache.maven.api.services.ModelBuilder;
+import org.apache.maven.api.services.ModelBuilderRequest;
+import org.apache.maven.api.services.ModelBuilderResult;
+import org.apache.maven.api.services.RepositoryFactory;
+import org.apache.maven.api.services.Sources;
 import org.apache.maven.cling.invoker.mvnup.UpgradeContext;
+import org.apache.maven.impl.standalone.ApiRunner;
+import org.codehaus.plexus.components.secdispatcher.Dispatcher;
+import org.codehaus.plexus.components.secdispatcher.internal.dispatchers.LegacyDispatcher;
+import org.eclipse.aether.spi.connector.transport.TransporterFactory;
+import org.eclipse.aether.spi.connector.transport.http.ChecksumExtractor;
+import org.eclipse.aether.spi.io.PathProcessor;
+import org.eclipse.aether.transport.file.FileTransporterFactory;
+import org.eclipse.aether.transport.jdk.JdkTransporterFactory;
 
 import static eu.maveniverse.domtrip.maven.MavenPomElements.Elements.PARENT;
 
@@ -46,6 +69,8 @@ import static eu.maveniverse.domtrip.maven.MavenPomElements.Elements.PARENT;
  * </pre>
  */
 public abstract class AbstractUpgradeStrategy implements UpgradeStrategy {
+
+    private Session session;
 
     /**
      * Template method that handles common logging and error handling.
@@ -184,5 +209,116 @@ public abstract class AbstractUpgradeStrategy implements UpgradeStrategy {
 
         context.info("Computed " + coordinatesByGAV.size() + " unique artifact(s) for inference");
         return new HashSet<>(coordinatesByGAV.values());
+    }
+
+    protected Session getSession() {
+        if (session == null) {
+            session = createMaven4Session();
+        }
+        return session;
+    }
+
+    private Session createMaven4Session() {
+        Session session = ApiRunner.createSession(injector -> {
+            injector.bindInstance(Dispatcher.class, new LegacyDispatcher());
+            injector.bindImplicit(TransporterFactoryConfig.class);
+        });
+
+        // TODO: we should read settings
+        RemoteRepository central =
+                session.createRemoteRepository(RemoteRepository.CENTRAL_ID, "https://repo.maven.apache.org/maven2");
+        RemoteRepository snapshots = session.getService(RepositoryFactory.class)
+                .createRemote(Repository.newBuilder()
+                        .id("apache-snapshots")
+                        .url("https://repository.apache.org/content/repositories/snapshots/")
+                        .releases(RepositoryPolicy.newBuilder().enabled("false").build())
+                        .snapshots(RepositoryPolicy.newBuilder().enabled("true").build())
+                        .build());
+
+        return session.withRemoteRepositories(List.of(central, snapshots));
+    }
+
+    protected Path createTempProjectStructure(UpgradeContext context, Map<Path, Document> pomMap) throws Exception {
+        Path tempDir = Files.createTempDirectory("mvnup-project-");
+        context.debug("Created temp project directory: " + tempDir);
+
+        Path commonRoot = findCommonRoot(pomMap.keySet());
+        context.debug("Common root: " + commonRoot);
+
+        for (Map.Entry<Path, Document> entry : pomMap.entrySet()) {
+            Path originalPath = entry.getKey();
+            Document document = entry.getValue();
+
+            Path relativePath = commonRoot.relativize(originalPath);
+            Path tempPomPath = tempDir.resolve(relativePath);
+
+            Files.createDirectories(tempPomPath.getParent());
+            Files.writeString(tempPomPath, document.toXml());
+            context.debug("Wrote POM to temp location: " + tempPomPath);
+        }
+
+        return tempDir;
+    }
+
+    protected Path findCommonRoot(Set<Path> pomPaths) {
+        Path commonRoot = null;
+        for (Path pomPath : pomPaths) {
+            Path parent = pomPath.getParent();
+            if (parent == null) {
+                parent = Path.of(".");
+            }
+            if (commonRoot == null) {
+                commonRoot = parent;
+            } else {
+                while (!parent.startsWith(commonRoot)) {
+                    commonRoot = commonRoot.getParent();
+                    if (commonRoot == null) {
+                        break;
+                    }
+                }
+            }
+        }
+        return commonRoot;
+    }
+
+    protected void cleanupTempDirectory(Path tempDir) {
+        try {
+            Files.walk(tempDir)
+                    .sorted(Comparator.reverseOrder())
+                    .map(Path::toFile)
+                    .forEach(File::delete);
+        } catch (Exception e) {
+            // Best effort cleanup
+        }
+    }
+
+    protected org.apache.maven.api.model.Model buildEffectiveModel(Path pomPath) {
+        Session session = getSession();
+        ModelBuilder modelBuilder = session.getService(ModelBuilder.class);
+
+        ModelBuilderRequest request = ModelBuilderRequest.builder()
+                .session(session)
+                .source(Sources.buildSource(pomPath))
+                .requestType(ModelBuilderRequest.RequestType.BUILD_EFFECTIVE)
+                .recursive(false)
+                .build();
+
+        ModelBuilderResult result = modelBuilder.newSession().build(request);
+        return result.getEffectiveModel();
+    }
+
+    static class TransporterFactoryConfig {
+        @Provides
+        @Named(JdkTransporterFactory.NAME)
+        static TransporterFactory jdkTransporterFactory(
+                ChecksumExtractor checksumExtractor, PathProcessor pathProcessor) {
+            return new JdkTransporterFactory(checksumExtractor, pathProcessor);
+        }
+
+        @Provides
+        @Named(FileTransporterFactory.NAME)
+        static TransporterFactory fileTransporterFactory() {
+            return new FileTransporterFactory();
+        }
     }
 }
